# Optimized `Rect/FRect` `pgRect_FromObject`